Tpetra,Ifpack2: Move equilibration test from Tpetra to Ifpack2
@trilinos/ifpack2 @trilinos/tpetra
