+++ packages/ifpack2/test/unit_tests/CMakeLists.txt
@@ -258,14 +258,3 @@ IF (${PACKAGE_NAME}_ENABLE_Experimental_KokkosKernels_Features)
     STANDARD_PASS_OUTPUT
     )
 ENDIF ()
-
-TRIBITS_ADD_EXECUTABLE_AND_TEST(
-  Equilibration
-  SOURCES 
-    Ifpack2_UnitTestEquilibration.cpp
-    ${TEUCHOS_STD_UNIT_TEST_MAIN}
-  ARGS ""
-  COMM serial mpi
-  NUM_MPI_PROCS 4
-  STANDARD_PASS_OUTPUT
-  )
