+++ packages/tpetra/core/test/CrsMatrix/CMakeLists.txt
@@ -321,6 +321,17 @@ TRIBITS_ADD_EXECUTABLE_AND_TEST(
   STANDARD_PASS_OUTPUT
   )
 
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  Equilibration
+  SOURCES 
+    Equilibration.cpp
+    ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  ARGS ""
+  COMM serial mpi
+  NUM_MPI_PROCS 4
+  STANDARD_PASS_OUTPUT
+  )
+
 SET(TIMING_INSTALLS "")
 
 INSTALL(TARGETS ${TIMING_INSTALLS}
